Results refactor (#73)
This PR refactored the results page. The lib we were using to fetch search results had quite a few limitations, so I refactored to use the NPM api more directly. this leads to a lot more code, but it gives us all the data we'll need to implement the a better results screen with information like download counts and package popularity